Fix CI/CD: Two more sign conversion issues for Clang-15
- lcode.cpp:668: Cast XOR result in SETARG_k
- parser.cpp:630: Cast getInfo() result in settablesize call

+++ src/compiler/lcode.cpp
@@ -665,7 +665,7 @@ void FuncState::negatecondition(expdesc *e) {
   Instruction *instr = getjumpcontrol(e->getInfo());
   lua_assert(testTMode(InstructionView(*instr).opcode()) && InstructionView(*instr).opcode() != OP_TESTSET &&
                                            InstructionView(*instr).opcode() != OP_TEST);
-  SETARG_k(*instr, (InstructionView(*instr).k() ^ 1));
+  SETARG_k(*instr, static_cast<unsigned int>(InstructionView(*instr).k() ^ 1));
 }
 
 /*
